logging configurable & fixed tests
