+++ magpie/magpie.ini
@@ -83,7 +83,7 @@ qualname = magpie
 [handler_console]
 class = StreamHandler
 args = (sys.stderr,)
-level = DEBUG
+level = INFO
 formatter = generic
 
 [formatter_generic]
